refactor: [Frontend] moved components into folders
