+++ frontend/src/components/mainpage/ApiStateIndicator.tsx
@@ -1,7 +1,7 @@
 import "./ApiStateIndicator.css"
 import axios from "axios";
 import {useEffect, useState} from "react";
-import {ApiState} from "../Types.tsx";
+import {ApiState} from "../../Types.tsx";
 
 export default function ApiStateIndicator() {
     const [ state, setState ] = useState<ApiState>()
