+++ frontend/src/components/mainpage/DarkModeSwitch.Functions.tsx
