+++ frontend/src/components/mainpage/SidePanel.css
