+++ frontend/src/components/mainpage/SidePanel.tsx
